Moved the AD file at the proper location
